Extracted interfaces for compiler and executor
